feat: add dynamic breadcrumbs navigation
