--- apps/frontend/app/components/marketplace/breadcrumb-navigation.tsx
@@ -1,24 +1,0 @@
-import {
-	Breadcrumb,
-	BreadcrumbItem,
-	BreadcrumbLink,
-	BreadcrumbList,
-	BreadcrumbPage,
-	BreadcrumbSeparator,
-} from "@/app/components/ui/breadcrumb";
-
-export default function BreadcrumbNavigation() {
-	return (
-		<Breadcrumb>
-			<BreadcrumbList className="text-base">
-				<BreadcrumbItem>
-					<BreadcrumbLink href="/">Home</BreadcrumbLink>
-				</BreadcrumbItem>
-				<BreadcrumbSeparator />
-				<BreadcrumbItem>
-					<BreadcrumbPage>Marketplace</BreadcrumbPage>
-				</BreadcrumbItem>
-			</BreadcrumbList>
-		</Breadcrumb>
-	);
-}
